Change how timers are defined
The issue was that the timers we started couldn't compute things at
the time they were triggered. e.g. you couldn't tell the user how much
time they have spent on the task. The flow was
  user functions -> funcall'd by us in a hook -> return values ->
  passed to `run-at-time' and `alert'
The arguments to `alert' were evaluated in the hook, rather than when
the notification was displayed.

Now each function calls run-at-time, instead of returning the
arguments for us to call run-at-time with.