Create sexp-open-log, rename open-file to open-log
The naming gave me trouble - 'file' isn't suitable for
database-backends. I considered 'view-' or 'edit-', but it is
currently meant for both.

We might be revisiting this, if users (currently just me?) - after an
SQLite backend is implemented - turn out to make frequent use of an
SQL shell to interact with the log.